Start code for recode.
Recode will rely on a simplified goal to reduce the code complexity.
Instead of an editable table heriarchy, only the top level table is editable
while the lower level tables are display only.

Backend system is for recode is nearly complete. Need to create a way to assign
ids to Tables and Table elements. the id counter should unique and persistent.

One idea is an IdAllocatotr class which is bound to the Project class. The Id Allocator
is responsible for stracking the next usable ID and incrementing the ID upon use.

View model for gui is still in works. Need to add an indexing operator to the table class
to get the viewmodel system to work. Also need to complete the gui interaction events
so everything properly binds.